[Example] fix some bugs
