+++ examples/pytorch/grand/README.md
@@ -49,14 +49,11 @@ The DGL's built-in Cora, Pubmed and Citeseer datasets. Dataset summary:
 --order            int     Propagation step.                      Default is 8.
 --sample           int     Sampling times of dropnode.            Default is 4.
 --tem              float   Sharpening temperaturer.               Default is 0.5.
---lam              float   Lamda.                                 Default is 1.0.
+--lam              float   Coefficient of Consistency reg         Default is 1.0.
 --use_bn           bool    Using batch normalization.             Default is False
 ```
 
 ## Examples
-
-The following commands learn a neural network and predict on the test set.
-Training a MixHop model on the default dataset.
 
 Train a model which follows the original hyperparameters on different datasets.
 ```bash
@@ -66,13 +63,17 @@ python main.py --dataname cora --gpu 0 --lam 1.0 --tem 0.5 --order 8 --sample 4 
 python main.py --dataname citeseer --gpu 0 --lam 0.7 --tem 0.3 --order 2 --sample 2 --input_droprate 0.0 --hidden_droprate 0.2 --dropnode_rate 0.5 --hid_dim 32 --early_stopping 100 --lr 1e-2  --epochs 2000
 
 # Pubmed:
-python main.py --dataname pubmed --gpu 0 --lam 1.0 --tem 0.2 --order 5 --sample 4 --input_droprate 0.6 --hidden_droprate 0.8 --dropnode_rate 0.5 --hid_dim 32 --early_stopping 100 --lr 0.2 --epochs 2000 --use_bn
+python main.py --dataname pubmed --gpu 0 --lam 1.0 --tem 0.2 --order 5 --sample 4 --input_droprate 0.6 --hidden_droprate 0.8 --dropnode_rate 0.5 --hid_dim 32 --early_stopping 200 --lr 0.2 --epochs 2000 --use_bn
 ```
 
 ### Performance
 
+The hyperparameter setting in our implementation is identical to that reported in the paper.
+
 | Dataset | Cora | Citeseer | Pubmed |
 | :-: | :-: | :-: | :-: |
 | Accuracy Reported(100 runs) | **85.4(±0.4)** | **75.4(±0.4)** | 82.7(±0.6) |
-| Accuracy DGL(20 runs) | **85.4(±0.4)** | 75.3(±0.3) | **82.9(±0.8)** |
+| Accuracy DGL(20 runs) | 85.33(±0.41) | 75.36(±0.36) | **82.90(±0.66)** |
+
+
