--- examples/pytorch/grand/layers.py
@@ -1,79 +1,0 @@
-import numpy as np
-import torch as th
-import torch.nn as nn
-from torch.nn import Linear, Parameter
-import torch.nn.functional as F
-import math
-
-
-def drop_node(feats, drop_rate, training):
-    
-    n = feats.shape[0]
-    drop_rates = th.FloatTensor(np.ones(n) * drop_rate)
-    
-    if training:
-            
-        masks = th.bernoulli(1. - drop_rates).unsqueeze(1)
-        feats = masks.cuda() * feats
-        
-    else:
-        feats = feats * (1. - drop_rate)
-
-    return feats
-    
-class MLPLayer(nn.Module):
-    def __init__(self, in_features, out_features, bias=True):
-        super(MLPLayer, self).__init__()
-        self.in_features = in_features
-        self.out_features = out_features
-        self.weight = Parameter(th.FloatTensor(in_features, out_features))
-        if bias:
-            self.bias = Parameter(th.FloatTensor(out_features))
-        else:
-            self.register_parameter('bias', None)
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        stdv = 1. / math.sqrt(self.weight.size(1))
-        self.weight.data.normal_(-stdv, stdv)
-        if self.bias is not None:
-            self.bias.data.normal_(-stdv, stdv)
-
-    def forward(self, input):
-        output = th.mm(input, self.weight)
-        if self.bias is not None:
-            return output + self.bias
-        else:
-            return output
-
-    def __repr__(self):
-        return self.__class__.__name__ + ' (' \
-               + str(self.in_features) + ' -> ' \
-               + str(self.out_features) + ')'    
-    
-class MLP(nn.Module):
-    def __init__(self, nfeat, nhid, nclass, input_droprate, hidden_droprate, use_bn =False):
-        super(MLP, self).__init__()
-
-        self.layer1 = MLPLayer(nfeat, nhid)
-        self.layer2 = MLPLayer(nhid, nclass)
-
-        self.input_droprate = input_droprate
-        self.hidden_droprate = hidden_droprate
-        self.bn1 = nn.BatchNorm1d(nfeat)
-        self.bn2 = nn.BatchNorm1d(nhid)
-        self.use_bn = use_bn
-        
-    def forward(self, x):
-         
-        if self.use_bn: 
-            x = self.bn1(x)
-        x = F.dropout(x, self.input_droprate, training=self.training)
-        x = F.relu(self.layer1(x))
-        if self.use_bn:
-            x = self.bn2(x)
-        x = F.dropout(x, self.hidden_droprate, training=self.training)
-        x = self.layer2(x)
-
-        return x
-
