+++ examples/pytorch/grand/model.py
@@ -1,8 +1,55 @@
+import numpy as np
 import torch as th
 import torch.nn as nn
 import dgl.function as fn
+import torch.nn.functional as F
 
-from layers import drop_node, MLP
+def drop_node(feats, drop_rate, training):
+    
+    n = feats.shape[0]
+    drop_rates = th.FloatTensor(np.ones(n) * drop_rate)
+    
+    if training:
+            
+        masks = th.bernoulli(1. - drop_rates).unsqueeze(1)
+        feats = masks.to(feats.device) * feats
+        
+    else:
+        feats = feats * (1. - drop_rate)
+
+    return feats
+
+class MLP(nn.Module):
+    def __init__(self, nfeat, nhid, nclass, input_droprate, hidden_droprate, use_bn =False):
+        super(MLP, self).__init__()
+        
+        self.layer1 = nn.Linear(nfeat, nhid, bias = True)
+        self.layer2 = nn.Linear(nhid, nclass, bias = True)
+
+        self.input_dropout = nn.Dropout(input_droprate)
+        self.hidden_dropout = nn.Dropout(hidden_droprate)
+        self.bn1 = nn.BatchNorm1d(nfeat)
+        self.bn2 = nn.BatchNorm1d(nhid)
+        self.use_bn = use_bn
+    
+    def reset_parameters(self):
+        self.layer1.reset_parameters()
+        self.layer2.reset_parameters()
+        
+    def forward(self, x):
+         
+        if self.use_bn: 
+            x = self.bn1(x)
+        x = self.input_dropout(x)
+        x = F.relu(self.layer1(x))
+        
+        if self.use_bn:
+            x = self.bn2(x)
+        x = self.hidden_dropout(x)
+        x = self.layer2(x)
+
+        return x   
+        
 
 def GRANDConv(graph, feats, order):
     '''
@@ -78,7 +125,7 @@ class GRAND(nn.Module):
         self.S = S
         self.K = K
         self.n_class = n_class
-
+        
         self.mlp = MLP(in_dim, hid_dim, n_class, input_droprate, hidden_droprate, batchnorm)
         
         self.dropout = node_dropout
@@ -102,3 +149,5 @@ class GRAND(nn.Module):
             X =  GRANDConv(graph, drop_feat, self.K)
 
             return th.log_softmax(self.mlp(X), dim = -1)
+        
+
